rename transition file
