+++ src/components/Layout.tsx
@@ -6,7 +6,7 @@ import Footer from "components/Footer"
 import Header from "components/Header"
 import Scroll from "components/Scroll"
 
-import { usePageLoad } from "./TransitionLink"
+import { usePageLoad } from "./TransitionUtils"
 
 interface LayoutProps {
   children: React.ReactNode
